libavfilter/qsvvpp: change the output frame's width and height
qsvvpp align the width and height with 16, and that right. But qsvvpp asign this value
to frame->width and frame->height, and that may lead to error.
For example, when we use qsvvpp to resize frame to 1080p, qsvvpp will
align frame to 1088 and set frame->height to 1088, which is different from the height of
encoder (1080) and this will be treated as resolution change. The aligend value can
be kept in hw_frame_ctx as it is hardware dependent. Now I
assign the out_link's w/h to output
frame to overwrite the w/h got from hw_frame_ctx.

Signed-off-by: Wenbin Chen <wenbin.chen@intel.com>